Rename factory package into instnaces to follow latest folder structuring
