+++ magnet-library/magnet-processor/src/main/java/magnet/processor/MagnetProcessor.kt
@@ -19,12 +19,12 @@ package magnet.processor
 import magnet.Instance
 import magnet.Magnetizer
 import magnet.Scope
-import magnet.processor.factory.CodeWriter
-import magnet.processor.factory.FactoryFromClassAnnotationParser
-import magnet.processor.factory.FactoryFromMethodAnnotationParser
-import magnet.processor.factory.FactoryIndexCodeGenerator
-import magnet.processor.factory.FactoryType
-import magnet.processor.factory.FactoryTypeCodeGenerator
+import magnet.processor.instances.CodeWriter
+import magnet.processor.instances.FactoryFromClassAnnotationParser
+import magnet.processor.instances.FactoryFromMethodAnnotationParser
+import magnet.processor.instances.FactoryIndexCodeGenerator
+import magnet.processor.instances.FactoryType
+import magnet.processor.instances.FactoryTypeCodeGenerator
 import magnet.processor.index.MagnetIndexerGenerator
 import magnet.processor.scopes.ScopeProcessor
 import javax.annotation.processing.AbstractProcessor
