+++ magnet-library/magnet-processor/src/main/java/magnet/processor/instances/CodeGenerator.kt
@@ -1,4 +1,4 @@
-package magnet.processor.factory
+package magnet.processor.instances
 
 interface CodeGenerator {
