+++ magnet-library/magnet-processor/src/main/java/magnet/processor/instances/FactoryFromMethodAnnotationParser.kt
@@ -1,4 +1,4 @@
-package magnet.processor.factory
+package magnet.processor.instances
 
 import com.squareup.javapoet.ClassName
 import magnet.Instance
