+++ magnet-library/magnet-processor/src/main/java/magnet/processor/instances/scoping/GetScopingMethodGenerator.kt
@@ -1,10 +1,10 @@
-package magnet.processor.factory.scoping
+package magnet.processor.instances.scoping
 
 import com.squareup.javapoet.MethodSpec
 import com.squareup.javapoet.TypeSpec
 import magnet.Scoping
-import magnet.processor.factory.GetScopingMethod
-import magnet.processor.factory.AspectGenerator
+import magnet.processor.instances.GetScopingMethod
+import magnet.processor.instances.AspectGenerator
 import javax.lang.model.element.Modifier
 
 internal class GetScopingMethodGenerator : AspectGenerator {
