+++ magnet-library/magnet-processor/src/main/java/magnet/processor/instances/siblingtypes/GetSiblingTypesMethodGenerator.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package magnet.processor.factory.siblingtypes
+package magnet.processor.instances.siblingtypes
 
 import com.squareup.javapoet.ArrayTypeName
 import com.squareup.javapoet.ClassName
@@ -22,8 +22,8 @@ import com.squareup.javapoet.CodeBlock
 import com.squareup.javapoet.FieldSpec
 import com.squareup.javapoet.MethodSpec
 import com.squareup.javapoet.TypeSpec
-import magnet.processor.factory.GetSiblingTypesMethod
-import magnet.processor.factory.AspectGenerator
+import magnet.processor.instances.GetSiblingTypesMethod
+import magnet.processor.instances.AspectGenerator
 import javax.lang.model.element.Modifier
 
 internal class GetSiblingTypesMethodGenerator : AspectGenerator {
