+++ magnet-library/magnet-processor/src/main/java/magnet/processor/scopes/instances/GetterMethodsGenerator.kt
@@ -14,7 +14,7 @@ internal class GetterMethodsGenerator : AspectGenerator() {
 
     private val builders = mutableListOf<MethodSpec.Builder>()
 
-    override fun enterScope(scope: Model.Scope) {
+    override fun visitScope(scope: Model.Scope) {
         builders.clear()
     }
