+++ magnet-library/magnet-processor/src/main/java/magnet/processor/scopes/instances/ScopeClassGenerator.kt
@@ -14,7 +14,7 @@ internal class ScopeClassGenerator : ClassGenerator() {
 
     override val packageName: String by lazy { scope.packageName }
 
-    override fun enterScope(scope: Model.Scope) {
+    override fun visitScope(scope: Model.Scope) {
         this.scope = scope
         classBuilder = TypeSpec.classBuilder(scope.getGeneratedScopeImplementationName())
             .addModifiers(Modifier.FINAL)
